+++ .vscode/settings.json
@@ -57,6 +57,7 @@
     "anonymised",
     "anonymiser",
     "Anonymising",
+    "anyio",
     "apipkg",
     "apputils",
     "arange",
